// CommerceLink/develop/VSI.EDGEAXConnector.SalesRepPortal.WebUI/src/app/_shared/models/address.ts
export class Address {
    Name: string;
    Street: string;
    City: string;
    ZipCode: string;
    country_code: string;
    Phone: string;
    AddressType: number;
    IsPrimary: boolean;
    RecordId:number;
    ThreeLetterISORegionName:string;
    EcomCustomerId:string;
    AddressTypeValue:number;
    AddressTypeStrValue:string;
    State:string;
    StreetNumber:string;
    TaxGroup:string;
    FullAddress: string;
}