// CommerceLink/develop/VSI.EDGEAXConnector.SalesRepPortal.WebUI/src/app/_shared/models/contact-person.ts
import { Address } from '../models/address';
export class ContactPerson {
    DirPartyRecordId: string;
    ContactPersonId: string;
    ContactForParty: string;
    CustAccount: string;
    InActive: boolean;
    Title: string;
    FirstName: string;
    MiddleName: string;
    LastName: string;
    EmailRecordId: string;
    Email: string;
    PhoneRecordId: string;
    Phone: string;
    Language: string;
    AddressName: string;
    Country: string;
    ZipCode: string;
    Street: string;
    City: string;
}